Refactoring scopes.
